refactor(cli): started code upgrade / overhaul
